Rename alertController to use a function presentBookmarkController
