+++ podcasts/Bookmarks/Podcast/BookmarksPodcastListController.swift
@@ -33,10 +33,6 @@ class BookmarksPodcastListController: ThemedHostingController<BookmarksPodcastLi
 // MARK: - BookmarkListRouter
 
 extension BookmarksPodcastListController: BookmarkListRouter {
-    var alertController: UIViewController? {
-        self
-    }
-
     func bookmarkPlay(_ bookmark: Bookmark) {
         playbackManager.playBookmark(bookmark)
         dismiss(animated: true)
